refactor lists into own files
